+++ src/app/product/[id]/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useSelectedOptions } from './useSelectedOptions';
